fix: [DEL-3845]: Move AcrBuildServiceImpl and related classes to 930-delegate-tasks.
