+++ 400-rest/src/main/java/software/wings/helpers/ext/azure/AzureHelperService.java
@@ -11,20 +11,14 @@ import static io.harness.annotations.dev.HarnessTeam.CDC;
 import static io.harness.beans.PageResponse.PageResponseBuilder.aPageResponse;
 import static io.harness.data.structure.EmptyPredicate.isEmpty;
 import static io.harness.data.structure.EmptyPredicate.isNotEmpty;
-import static io.harness.eraro.ErrorCode.AZURE_SERVICE_EXCEPTION;
 import static io.harness.eraro.ErrorCode.INVALID_AZURE_VAULT_CONFIGURATION;
 import static io.harness.exception.WingsException.USER;
-import static io.harness.network.Http.getOkHttpClientBuilder;
 import static io.harness.validation.Validator.notNullCheck;
 
 import static software.wings.beans.infrastructure.Host.Builder.aHost;
 
-import static com.google.common.base.Charsets.UTF_8;
 import static com.microsoft.azure.management.compute.PowerState.RUNNING;
-import static java.lang.String.format;
 import static java.util.stream.Collectors.toList;
-import static java.util.stream.Collectors.toSet;
-import static org.apache.commons.codec.binary.Base64.encodeBase64String;
 
 import io.harness.annotations.dev.HarnessModule;
 import io.harness.annotations.dev.OwnedBy;
@@ -35,19 +29,13 @@ import io.harness.exception.AzureServiceException;
 import io.harness.exception.ExceptionUtils;
 import io.harness.exception.InvalidArgumentsException;
 import io.harness.exception.InvalidRequestException;
-import io.harness.network.Http;
 import io.harness.security.encryption.EncryptedDataDetail;
 
 import software.wings.api.DeploymentType;
 import software.wings.beans.AzureAvailabilitySet;
 import software.wings.beans.AzureConfig;
-import software.wings.beans.AzureContainerRegistry;
-import software.wings.beans.AzureImageDefinition;
-import software.wings.beans.AzureImageGallery;
-import software.wings.beans.AzureImageVersion;
 import software.wings.beans.AzureInfrastructureMapping;
 import software.wings.beans.AzureTag;
-import software.wings.beans.AzureTagDetails;
 import software.wings.beans.AzureVaultConfig;
 import software.wings.beans.AzureVirtualMachineScaleSet;
 import software.wings.beans.SettingAttribute;
@@ -65,12 +53,9 @@ import com.microsoft.azure.credentials.ApplicationTokenCredentials;
 import com.microsoft.azure.management.Azure;
 import com.microsoft.azure.management.Azure.Authenticated;
 import com.microsoft.azure.management.compute.VirtualMachine;
-import com.microsoft.azure.management.containerregistry.Registry;
 import com.microsoft.azure.management.containerservice.OSType;
 import com.microsoft.azure.management.keyvault.Vault;
 import com.microsoft.azure.management.resources.ResourceGroup;
-import com.microsoft.azure.management.resources.Subscription;
-import com.microsoft.azure.management.resources.fluentcore.arm.models.HasName;
 import com.microsoft.rest.LogLevel;
 import java.io.IOException;
 import java.util.ArrayList;
@@ -78,15 +63,9 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
-import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 import lombok.extern.slf4j.Slf4j;
-import okhttp3.OkHttpClient;
 import org.apache.commons.lang3.tuple.ImmutablePair;
-import retrofit2.Response;
-import retrofit2.Retrofit;
-import retrofit2.converter.jackson.JacksonConverterFactory;
 
 @OwnedBy(CDC)
 @Singleton
@@ -139,23 +118,6 @@ public class AzureHelperService {
     } catch (Exception e) {
       handleAzureAuthenticationException(e);
     }
-  }
-
-  public Map<String, String> listSubscriptions(AzureConfig azureConfig, List<EncryptedDataDetail> encryptionDetails) {
-    encryptionService.decrypt(azureConfig, encryptionDetails, false);
-    Azure azure = getAzureClient(azureConfig);
-    return azure.subscriptions().list().stream().collect(
-        Collectors.toMap(Subscription::subscriptionId, Subscription::displayName));
-  }
-
-  public boolean isValidSubscription(AzureConfig azureConfig, String subscriptionId) {
-    return getAzureClient(azureConfig)
-               .subscriptions()
-               .list()
-               .stream()
-               .filter(subscription -> subscription.subscriptionId().equalsIgnoreCase(subscriptionId))
-               .count()
-        != 0;
   }
 
   public List<VirtualMachine> listVms(AzureInfrastructureMapping azureInfrastructureMapping,
@@ -363,229 +325,6 @@ public class AzureHelperService {
         .collect(toList());
   }
 
-  public List<AzureTagDetails> listTags(
-      AzureConfig azureConfig, List<EncryptedDataDetail> encryptionDetails, String subscriptionId) {
-    encryptionService.decrypt(azureConfig, encryptionDetails, false);
-    try {
-      Response<AzureListTagsResponse> response = getAzureManagementRestClient(azureConfig.getAzureEnvironmentType())
-                                                     .listTags(getAzureBearerAuthToken(azureConfig), subscriptionId)
-                                                     .execute();
-
-      if (response.isSuccessful()) {
-        return response.body()
-            .getValue()
-            .stream()
-            .map(tagDetails
-                -> AzureTagDetails.builder()
-                       .tagName(tagDetails.getTagName())
-                       .values(tagDetails.getValues().stream().map(TagValue::getTagValue).collect(toList()))
-                       .build())
-            .collect(toList());
-      } else {
-        log.error("Error occurred while getting Tags from subscriptionId : " + subscriptionId
-            + " Response: " + response.raw());
-        throw new AzureServiceException(response.message(), AZURE_SERVICE_EXCEPTION, USER);
-      }
-    } catch (Exception e) {
-      handleAzureAuthenticationException(e);
-      return null;
-    }
-  }
-
-  public Set<String> listTagsBySubscription(
-      String subscriptionId, AzureConfig azureConfig, List<EncryptedDataDetail> encryptionDetails) {
-    encryptionService.decrypt(azureConfig, encryptionDetails, false);
-    try {
-      Response<AzureListTagsResponse> response = getAzureManagementRestClient(azureConfig.getAzureEnvironmentType())
-                                                     .listTags(getAzureBearerAuthToken(azureConfig), subscriptionId)
-                                                     .execute();
-
-      if (response.isSuccessful()) {
-        return response.body().getValue().stream().map(TagDetails::getTagName).collect(toSet());
-      } else {
-        log.error("Error occurred while getting Tags from subscriptionId : " + subscriptionId
-            + " Response: " + response.raw());
-        throw new AzureServiceException(response.message(), AZURE_SERVICE_EXCEPTION, USER);
-      }
-    } catch (Exception e) {
-      handleAzureAuthenticationException(e);
-      return null;
-    }
-  }
-
-  public Set<String> listResourceGroups(
-      AzureConfig azureConfig, List<EncryptedDataDetail> encryptionDetails, String subscriptionId) {
-    encryptionService.decrypt(azureConfig, encryptionDetails, false);
-
-    try {
-      Azure azure = getAzureClient(azureConfig, subscriptionId);
-      notNullCheck("Azure Client", azure);
-      List<ResourceGroup> resourceGroupList = azure.resourceGroups().list();
-      return resourceGroupList.stream().map(HasName::name).collect(Collectors.toSet());
-    } catch (Exception e) {
-      handleAzureAuthenticationException(e);
-    }
-    return Collections.EMPTY_SET;
-  }
-
-  public List<String> listContainerRegistryNames(AzureConfig azureConfig, String subscriptionId) {
-    Azure azure = getAzureClient(azureConfig, subscriptionId);
-    List<String> registries = new ArrayList<>();
-    azure.containerRegistries().list().forEach(registry -> registries.add(registry.name()));
-    return registries;
-  }
-
-  public List<AzureContainerRegistry> listContainerRegistries(AzureConfig azureConfig, String subscriptionId) {
-    Azure azure = getAzureClient(azureConfig, subscriptionId);
-    List<AzureContainerRegistry> registries = new ArrayList<>();
-    azure.containerRegistries().list().forEach(registry
-        -> registries.add(AzureContainerRegistry.builder()
-                              .id(registry.id())
-                              .name(registry.name())
-                              .resourceGroup(registry.resourceGroupName())
-                              .subscriptionId(subscriptionId)
-                              .loginServer(registry.loginServerUrl())
-                              .build()));
-    return registries;
-  }
-
-  public boolean isValidContainerRegistry(AzureConfig azureConfig, String subscriptionId, String registryName) {
-    return getAzureClient(azureConfig, subscriptionId)
-               .containerRegistries()
-               .list()
-               .stream()
-               .filter(registry -> registry.name().equalsIgnoreCase(registryName))
-               .count()
-        != 0;
-  }
-
-  public String getLoginServerForRegistry(AzureConfig azureConfig, List<EncryptedDataDetail> encryptionDetails,
-      String subscriptionId, String registryName) {
-    encryptionService.decrypt(azureConfig, encryptionDetails, false);
-    return getRegistry(azureConfig, encryptionDetails, subscriptionId, registryName).loginServerUrl();
-  }
-
-  private Registry getRegistry(AzureConfig azureConfig, List<EncryptedDataDetail> encryptionDetails,
-      String subscriptionId, String registryName) {
-    encryptionService.decrypt(azureConfig, encryptionDetails, false);
-    return getAzureClient(azureConfig, subscriptionId)
-        .containerRegistries()
-        .list()
-        .stream()
-        .filter(item -> item.name().equals(registryName))
-        .findFirst()
-        .get();
-  }
-
-  public List<String> listRepositories(AzureConfig azureConfig, String subscriptionId, String registryName) {
-    Azure azure = getAzureClient(azureConfig, subscriptionId);
-    try {
-      Registry registry = azure.containerRegistries()
-                              .list()
-                              .stream()
-                              .filter(item -> item.name().equals(registryName))
-                              .findFirst()
-                              .get();
-      AcrRestClient acrRestClient = getAcrRestClient(registry.loginServerUrl());
-      List<String> allRepositories = new ArrayList<>();
-      String last = null;
-      List<String> repositories;
-      do {
-        repositories =
-            acrRestClient
-                .listRepositories(getAuthHeader(azureConfig.getClientId(), new String(azureConfig.getKey())), last)
-                .execute()
-                .body()
-                .getRepositories();
-        if (isNotEmpty(repositories)) {
-          allRepositories.addAll(repositories);
-          last = repositories.get(repositories.size() - 1);
-        }
-      } while (isNotEmpty(repositories));
-      return allRepositories.stream().distinct().collect(toList());
-    } catch (Exception e) {
-      log.error("Error occurred while getting repositories from subscriptionId/registryName :" + subscriptionId + "/"
-              + registryName,
-          e);
-      throw new AzureServiceException(
-          "Failed to list repositories " + ExceptionUtils.getMessage(e), AZURE_SERVICE_EXCEPTION, USER);
-    }
-  }
-
-  public List<String> listRepositoryTags(AzureConfig azureConfig, List<EncryptedDataDetail> encryptionDetails,
-      String subscriptionId, String registryName, String repositoryName) {
-    encryptionService.decrypt(azureConfig, encryptionDetails, false);
-    Azure azure = getAzureClient(azureConfig, subscriptionId);
-    try {
-      Registry registry = azure.containerRegistries()
-                              .list()
-                              .stream()
-                              .filter(item -> item.name().equals(registryName))
-                              .findFirst()
-                              .get();
-      AcrRestClient acrRestClient = getAcrRestClient(registry.loginServerUrl());
-      return acrRestClient
-          .listRepositoryTags(
-              getAuthHeader(azureConfig.getClientId(), new String(azureConfig.getKey())), repositoryName)
-          .execute()
-          .body()
-          .getTags();
-    } catch (Exception e) {
-      log.error("Error occurred while getting repositories from subscriptionId/registryName/repositoryName :"
-              + subscriptionId + "/" + registryName + "/" + repositoryName,
-          e);
-      throw new AzureServiceException(
-          "Failed to retrieve repository tags " + ExceptionUtils.getMessage(e), AZURE_SERVICE_EXCEPTION, USER);
-    }
-  }
-
-  public List<String> listRepositoryTags(AzureConfig azureConfig, List<EncryptedDataDetail> encryptionDetails,
-      String registryHostName, String repositoryName) {
-    encryptionService.decrypt(azureConfig, encryptionDetails, false);
-    try {
-      AcrRestClient acrRestClient = getAcrRestClient(registryHostName);
-      return acrRestClient
-          .listRepositoryTags(
-              getAuthHeader(azureConfig.getClientId(), new String(azureConfig.getKey())), repositoryName)
-          .execute()
-          .body()
-          .getTags();
-    } catch (Exception e) {
-      log.error("Error occurred while getting Tags for Repository :" + registryHostName + "/" + repositoryName, e);
-      throw new AzureServiceException(
-          "Failed to retrieve repository tags " + ExceptionUtils.getMessage(e), AZURE_SERVICE_EXCEPTION, USER);
-    }
-  }
-
-  @VisibleForTesting
-  String getAzureBearerAuthToken(AzureConfig azureConfig) {
-    try {
-      AzureEnvironment azureEnvironment = getAzureEnvironment(azureConfig.getAzureEnvironmentType());
-      ApplicationTokenCredentials credentials = new ApplicationTokenCredentials(
-          azureConfig.getClientId(), azureConfig.getTenantId(), new String(azureConfig.getKey()), azureEnvironment);
-
-      String token = credentials.getToken(azureEnvironment.managementEndpoint());
-      return "Bearer " + token;
-    } catch (Exception e) {
-      handleAzureAuthenticationException(e);
-    }
-    return null;
-  }
-
-  @VisibleForTesting
-  protected Azure getAzureClient(AzureConfig azureConfig) {
-    try {
-      ApplicationTokenCredentials credentials =
-          new ApplicationTokenCredentials(azureConfig.getClientId(), azureConfig.getTenantId(),
-              new String(azureConfig.getKey()), getAzureEnvironment(azureConfig.getAzureEnvironmentType()));
-
-      return Azure.configure().withLogLevel(LogLevel.NONE).authenticate(credentials).withDefaultSubscription();
-    } catch (Exception e) {
-      handleAzureAuthenticationException(e);
-    }
-    return null;
-  }
-
   @VisibleForTesting
   protected Azure getAzureClient(AzureConfig azureConfig, String subscriptionId) {
     try {
@@ -598,47 +337,6 @@ public class AzureHelperService {
       handleAzureAuthenticationException(e);
     }
     return null;
-  }
-
-  private AcrRestClient getAcrRestClient(String registryHostName) {
-    String url = getUrl(registryHostName);
-    OkHttpClient okHttpClient = getOkHttpClientBuilder()
-                                    .connectTimeout(CONNECT_TIMEOUT, TimeUnit.SECONDS)
-                                    .readTimeout(READ_TIMEOUT, TimeUnit.SECONDS)
-                                    .proxy(Http.checkAndGetNonProxyIfApplicable(url))
-                                    .retryOnConnectionFailure(true)
-                                    .build();
-    Retrofit retrofit = new Retrofit.Builder()
-                            .client(okHttpClient)
-                            .baseUrl(url)
-                            .addConverterFactory(JacksonConverterFactory.create())
-                            .build();
-    return retrofit.create(AcrRestClient.class);
-  }
-
-  @VisibleForTesting
-  AzureManagementRestClient getAzureManagementRestClient(AzureEnvironmentType azureEnvironmentType) {
-    String url = getAzureEnvironment(azureEnvironmentType).resourceManagerEndpoint();
-    OkHttpClient okHttpClient = getOkHttpClientBuilder()
-                                    .connectTimeout(CONNECT_TIMEOUT, TimeUnit.SECONDS)
-                                    .readTimeout(READ_TIMEOUT, TimeUnit.SECONDS)
-                                    .proxy(Http.checkAndGetNonProxyIfApplicable(url))
-                                    .retryOnConnectionFailure(true)
-                                    .build();
-    Retrofit retrofit = new Retrofit.Builder()
-                            .client(okHttpClient)
-                            .baseUrl(url)
-                            .addConverterFactory(JacksonConverterFactory.create())
-                            .build();
-    return retrofit.create(AzureManagementRestClient.class);
-  }
-
-  private String getAuthHeader(String username, String password) {
-    return "Basic " + encodeBase64String(format("%s:%s", username, password).getBytes(UTF_8));
-  }
-
-  public String getUrl(String acrHostName) {
-    return "https://" + acrHostName + (acrHostName.endsWith("/") ? "" : "/");
   }
 
   public void handleAzureAuthenticationException(Exception e) {
@@ -662,65 +360,6 @@ public class AzureHelperService {
       log.error("Listing vaults failed for account Id {}", accountId, ex);
       throw new AzureServiceException("Failed to list vaults.", INVALID_AZURE_VAULT_CONFIGURATION, USER);
     }
-  }
-
-  public List<AzureImageGallery> listImageGalleries(AzureConfig azureConfig,
-      List<EncryptedDataDetail> encryptionDetails, String subscriptionId, String resourceGroupName) {
-    encryptionService.decrypt(azureConfig, encryptionDetails, false);
-    Azure azure = getAzureClient(azureConfig, subscriptionId);
-    return azure.galleries()
-        .listByResourceGroup(resourceGroupName)
-        .stream()
-        .map(ig
-            -> AzureImageGallery.builder()
-                   .name(ig.name())
-                   .subscriptionId(subscriptionId)
-                   .resourceGroupName(resourceGroupName)
-                   .regionName(ig.regionName())
-                   .build())
-        .collect(Collectors.toList());
-  }
-
-  public List<AzureImageDefinition> listImageDefinitions(AzureConfig azureConfig,
-      List<EncryptedDataDetail> encryptionDetails, String subscriptionId, String resourceGroupName,
-      String galleryName) {
-    encryptionService.decrypt(azureConfig, encryptionDetails, false);
-    Azure azure = getAzureClient(azureConfig, subscriptionId);
-    return azure.galleryImages()
-        .listByGallery(resourceGroupName, galleryName)
-        .stream()
-        .map(id
-            -> AzureImageDefinition.builder()
-                   .name(id.name())
-                   .location(id.location())
-                   .osType(id.osType().name())
-                   .subscriptionId(subscriptionId)
-                   .resourceGroupName(resourceGroupName)
-                   .galleryName(galleryName)
-                   .build())
-        .collect(Collectors.toList());
-  }
-
-  public List<AzureImageVersion> listImageDefinitionVersions(AzureConfig azureConfig,
-      List<EncryptedDataDetail> encryptionDetails, String subscriptionId, String resourceGroupName, String galleryName,
-      String imageDefinition) {
-    encryptionService.decrypt(azureConfig, encryptionDetails, false);
-    Azure azure = getAzureClient(azureConfig, subscriptionId);
-    // Only fetch successful Azure Image versions
-    return azure.galleryImageVersions()
-        .listByGalleryImage(resourceGroupName, galleryName, imageDefinition)
-        .stream()
-        .filter(id -> "Succeeded".equals(id.provisioningState()))
-        .map(id
-            -> AzureImageVersion.builder()
-                   .name(id.name())
-                   .imageDefinitionName(imageDefinition)
-                   .location(id.location())
-                   .subscriptionId(subscriptionId)
-                   .resourceGroupName(resourceGroupName)
-                   .galleryName(galleryName)
-                   .build())
-        .collect(Collectors.toList());
   }
 
   private List<Vault> listVaultsInternal(String accountId, AzureVaultConfig azureVaultConfig) throws IOException {
