+++ 960-api-services/src/main/java/software/wings/helpers/ext/azure/AcrRestClient.java
