+++ 970-api-services-beans/src/main/java/software/wings/beans/AzureTagDetails.java
